requirements-analysis.txt
pandas
sqlalchemy
psycopg[binary]
jupyter
matplotlib
seaborn
python-dotenv
